Move UTC-OFFSET type to new data type system
