+++ ical/types/__init__.py
@@ -1,4 +1,4 @@
-"""Library for parsing rfc5545 Property Value Data Types."""
+"""Library for parsing rfc5545 Property Value Data Types and Properties."""
 
 # Import all types for the registry
 from . import integer  # noqa: F401
@@ -6,5 +6,6 @@ from . import boolean, date, date_time, duration  # noqa: F401
 from . import float as float_pkg  # noqa: F401
 from .geo import Geo
 from .period import Period
+from .utc_offset import UtcOffset
 
-__all__ = ["Period", "Geo"]
+__all__ = ["Period", "Geo", "UtcOffset"]
