+++ tests/test_types.py
@@ -6,10 +6,9 @@ from typing import Optional, Union
 import pytest
 from pydantic import ValidationError
 
-from ical._types import ICS_ENCODERS, ComponentModel, Priority, UtcOffset
+from ical._types import ICS_ENCODERS, ComponentModel, Priority
 from ical.parsing.component import ParsedComponent
 from ical.parsing.property import ParsedProperty
-from ical.types import Period
 
 
 def test_encode_component() -> None:
@@ -158,271 +157,3 @@ def test_priority() -> None:
 
     with pytest.raises(ValidationError):
         TestModel.parse_obj({"pri": [ParsedProperty(name="dt", value="10")]})
-
-
-def test_bool() -> None:
-    """Test for boolean fields."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        example: bool
-
-    model = TestModel.parse_obj(
-        {"example": [ParsedProperty(name="example", value="TRUE")]}
-    )
-    assert model.example
-
-    model = TestModel.parse_obj(
-        {"example": [ParsedProperty(name="example", value="FALSE")]}
-    )
-    assert not model.example
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj({"example": [ParsedProperty(name="example", value="efd")]})
-
-    # Populate based on bool object
-    model = TestModel(example=True)
-    assert model.example
-    component = model.__encode_component_root__()
-    assert component.properties == [
-        ParsedProperty(name="example", value="TRUE"),
-    ]
-
-    model = TestModel(example=False)
-    assert not model.example
-    component = model.__encode_component_root__()
-    assert component.properties == [
-        ParsedProperty(name="example", value="FALSE"),
-    ]
-
-
-@pytest.mark.parametrize(
-    "value,duration,encoded_value",
-    [
-        (
-            "P15DT5H0M20S",
-            datetime.timedelta(days=15, hours=5, seconds=20),
-            "P2W1DT5H20S",
-        ),
-        ("P7W", datetime.timedelta(days=7 * 7), "P7W"),
-        ("-P7W", datetime.timedelta(days=-7 * 7), "-P7W"),
-    ],
-)
-def test_duration(value: str, duration: datetime.timedelta, encoded_value: str) -> None:
-    """Test for duration fields."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        duration: datetime.timedelta
-
-        class Config:
-            """Pydantic model configuration."""
-
-            json_encoders = ICS_ENCODERS
-
-    model = TestModel.parse_obj(
-        {"duration": [ParsedProperty(name="duration", value=value)]}
-    )
-    assert model.duration == duration
-    component = model.__encode_component_root__()
-    assert component.name == "TestModel"
-    assert component.properties == [
-        ParsedProperty(name="duration", value=encoded_value)
-    ]
-
-
-def test_duration_from_object() -> None:
-    """Test for a duration field from a native object."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        duration: datetime.timedelta
-
-        class Config:
-            """Pydantic model configuration."""
-
-            json_encoders = ICS_ENCODERS
-
-    model = TestModel(duration=datetime.timedelta(hours=1))
-    assert model.duration == datetime.timedelta(hours=1)
-
-    component = model.__encode_component_root__()
-    assert component.name == "TestModel"
-    assert component.properties == [ParsedProperty(name="duration", value="PT1H")]
-
-
-def test_integer() -> None:
-    """Test for int fields."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        example: list[int]
-
-    model = TestModel.parse_obj(
-        {
-            "example": [
-                ParsedProperty(name="example", value="45"),
-                ParsedProperty(name="example", value="-46"),
-                ParsedProperty(name="example", value="+47"),
-            ]
-        }
-    )
-    assert model.example == [45, -46, 47]
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj({"example": [ParsedProperty(name="example", value="a")]})
-
-
-def test_float() -> None:
-    """Test for float fields."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        example: list[float]
-
-    model = TestModel.parse_obj(
-        {
-            "example": [
-                ParsedProperty(name="example", value="45"),
-                ParsedProperty(name="example", value="-46.2"),
-                ParsedProperty(name="example", value="+47.32"),
-            ]
-        }
-    )
-    assert model.example == [45, -46.2, 47.32]
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj({"example": [ParsedProperty(name="example", value="a")]})
-
-    model = TestModel(example=[1, -2.2, 3.5])
-    assert model.example == [1, -2.2, 3.5]
-
-
-def test_period() -> None:
-    """Test for period fields."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        example: Period
-
-    # Time period with end date
-    model = TestModel.parse_obj(
-        {
-            "example": [
-                ParsedProperty(
-                    name="example", value="19970101T180000Z/19970102T070000Z"
-                )
-            ]
-        },
-    )
-    assert model.example.start == datetime.datetime(
-        1997, 1, 1, 18, 0, 0, tzinfo=datetime.timezone.utc
-    )
-    assert model.example.end
-    assert not model.example.duration
-    assert model.example.end_value == datetime.datetime(
-        1997, 1, 2, 7, 0, 0, tzinfo=datetime.timezone.utc
-    )
-
-    # Time period with duration
-    model = TestModel.parse_obj(
-        {"example": [ParsedProperty(name="example", value="19970101T180000Z/PT5H30M")]},
-    )
-    assert model.example.start == datetime.datetime(
-        1997, 1, 1, 18, 0, 0, tzinfo=datetime.timezone.utc
-    )
-
-    assert not model.example.end
-    assert model.example.duration
-    assert model.example.end_value == datetime.datetime(
-        1997, 1, 1, 23, 30, 0, tzinfo=datetime.timezone.utc
-    )
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj({"example": [ParsedProperty(name="example", value="a")]})
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj(
-            {"example": [ParsedProperty(name="example", value="19970101T180000Z/a")]}
-        )
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj(
-            {"example": [ParsedProperty(name="example", value="a/19970102T070000Z")]}
-        )
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj(
-            {"example": [ParsedProperty(name="example", value="a/PT5H30M")]}
-        )
-
-
-def test_encode_period() -> None:
-    """Test encoded period."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        example: Period
-
-        class Config:
-            """Pydantic model configuration."""
-
-            json_encoders = ICS_ENCODERS
-
-    model = TestModel(
-        example=Period(
-            start=datetime.datetime(2022, 8, 7, 6, 0, 0),
-            end=datetime.datetime(2022, 8, 7, 6, 30, 0),
-        )
-    )
-    assert model.__encode_component_root__() == ParsedComponent(
-        name="TestModel",
-        properties=[
-            ParsedProperty(name="example", value="20220807T060000/20220807T063000")
-        ],
-    )
-
-    model = TestModel(
-        example=Period(
-            start=datetime.datetime(2022, 8, 7, 6, 0, 0),
-            duration=datetime.timedelta(hours=5, minutes=30),
-        )
-    )
-    assert model.__encode_component_root__() == ParsedComponent(
-        name="TestModel",
-        properties=[ParsedProperty(name="example", value="20220807T060000/PT5H30M")],
-    )
-
-
-def test_utc_offset() -> None:
-    """Test for UTC offset fields."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        example: UtcOffset
-
-    model = TestModel.parse_obj(
-        {"example": [ParsedProperty(name="example", value="-0400")]}
-    )
-    assert model.example.offset == datetime.timedelta(hours=-4)
-
-    model = TestModel.parse_obj(
-        {"example": [ParsedProperty(name="example", value="0500")]}
-    )
-    assert model.example.offset == datetime.timedelta(hours=5)
-
-    model = TestModel(example=UtcOffset(offset=datetime.timedelta(hours=5)))
-    assert model.example.offset == datetime.timedelta(hours=5)
-
-    with pytest.raises(ValidationError, match=r".*match UTC-OFFSET pattern.*"):
-        TestModel.parse_obj(
-            {"example": [ParsedProperty(name="example", value="abcdef")]},
-        )
